# backend/openapi.yaml
openapi: 3.0.3
info:
  title: PawRoom Backend MVP API
  version: 0.1.0
  description: Mock collar telemetry, pet state snapshots, realtime scene events, and one-off memory creation.
servers:
  - url: http://localhost:4000
paths:
...
  /pets/{petId}/timeline:
    get:
      summary: Get timeline events for a pet.
      parameters:
        - name: petId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Timeline events.
...